@@ -1,18 +1,17 @@
 package com.phantomvk.democenter
 
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
+import androidx.databinding.DataBindingUtil
 import com.phantomvk.democenter.activity.BubbleShapeActivity
+import com.phantomvk.democenter.databinding.ActivityMainBinding
 import com.phantomvk.democenter.util.startActivity
-import kotlinx.android.synthetic.main.activity_main.*
 
 class MainActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-
-        // To BubbleShapeActivity.
-        btnBubbleShape.startActivity<BubbleShapeActivity>()
+        val binding = DataBindingUtil.setContentView<ActivityMainBinding>(this, R.layout.activity_main)
+        binding.btnBubbleShape.startActivity<BubbleShapeActivity>()
     }
 }
